Fix: Change Post method from Customer to Order
