+++ webapi/Controllers/CustomersController.cs
@@ -22,7 +22,7 @@ namespace webapi.Controllers
           var customers = _unitOfWork.Customers.GetAll();
           var totalCount = customers.Count<Customer> ();
 
-          return Ok (new { success = true, data = customers });
+          return Ok (new { success = true, data = customers, totalCount = totalCount });
         }
 
         // GET: api/customers/id
@@ -57,98 +57,6 @@ namespace webapi.Controllers
           _unitOfWork.Complete();
 
           return Ok (new { success = true, data = customer });
-        }
-
-        // POST: api/customers
-        [HttpPost]
-        public ActionResult PostCustomer(AddCustomer values) {
-          var customer = _unitOfWork.Customers.GetBy(values.CustomerId);
-
-          if (customer != null) {
-            customer.BookingCount++;
-          } else {
-            // Add Customer
-            customer = new Customer { 
-              Id = values.CustomerId,
-              FullName = values.FullName,
-              Phone = values.Phone,
-              BookingCount = values.BookingCount
-            };
-
-            _unitOfWork.Customers.Add(customer);
-          }
-
-          // Add Order
-          Order order = new Order {
-            Id = this.autoOrderId(),
-            TicketCount = values.TicketCount,
-            TotalPrice = values.TotalPrice,
-            CreateAt = values.CreateAt,
-            Status = values.Status,
-            CustomerId = values.CustomerId,
-          };
-    
-          _unitOfWork.Orders.Add(order);
-          _unitOfWork.Complete();
-    
-          return Ok (new { success = true, message = "Add Successfully" });
-        }
-        
-        // DELETE: api/customers/id
-        [HttpDelete ("{id}")]
-        public ActionResult DeleteCustomer(string id) {
-          var customer = _unitOfWork.Customers.GetBy(id);
-          
-          if (customer == null) {
-            return NotFound (new { success = false, message = "Invalid Customer" });
-          }
-
-          _unitOfWork.Customers.Remove(customer);
-          _unitOfWork.Complete();
-
-          return Ok (new { success = true, message = "Delete Successfully" });
-        }
-        
-        // Hàm phát sinh:
-        // 0. Kiểm tra có bao nhiêu chữ số
-        private int totalDigits(int number) {
-          int sum = 0;
-
-          while (number != 0) {
-            sum++;
-            number = number / 10;
-          }
-
-          return sum;
-        }
-
-        // 1. Tự động sinh OrderId
-        private string autoOrderId() {
-          string orderId = "";
-          var orders = _unitOfWork.Orders.GetAll();
-
-          if (orders.Any()) {
-            int orderIdNum = Int32.Parse(orders.Last().Id) + 1; // Lấy mã đơn hàng cũ chuyển qua kiểu int và + 1
-            int totalDigits = this.totalDigits(orderIdNum); // Tính tổng chữ số của mã mới lấy được
-            switch(totalDigits) {
-              case 1:
-                orderId = "000" + orderIdNum;
-                break;
-              case 2:
-                orderId = "00" + orderIdNum;
-                break;
-              case 3:
-                orderId = "0" + orderIdNum;
-                break;
-              default:
-                orderId += orderIdNum;
-                break;
-            }
-          } else {
-            orderId = "0001";
-          }
-
-          return orderId;
         }
     }
 }
